Item and Rent adjustments and adding custom exceptions and http status

+++ REST/src/main/java/pl/lodz/p/edu/rest/service/ItemService.java
@@ -8,9 +8,14 @@ import pl.lodz.p.edu.rest.dto.ComicsDTO;
 import pl.lodz.p.edu.rest.dto.ItemDTO;
 import pl.lodz.p.edu.rest.dto.MovieDTO;
 import pl.lodz.p.edu.rest.dto.MusicDTO;
+import pl.lodz.p.edu.rest.exception.InvalidItemTypeException;
+import pl.lodz.p.edu.rest.exception.ItemAlreadyRentedException;
+import pl.lodz.p.edu.rest.exception.ItemNotFoundException;
 import pl.lodz.p.edu.rest.mapper.ItemMapper;
+import pl.lodz.p.edu.rest.model.Rent;
 import pl.lodz.p.edu.rest.model.item.*;
 import pl.lodz.p.edu.rest.repository.ItemRepository;
+import pl.lodz.p.edu.rest.repository.RentRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -19,10 +24,12 @@ import java.util.stream.Collectors;
 @Service
 public class ItemService {
     private final ItemRepository itemRepository;
+    private final RentRepository rentRepository;
 
     @Autowired
-    public ItemService(ItemRepository itemRepository) {
+    public ItemService(ItemRepository itemRepository, RentRepository rentRepository) {
         this.itemRepository = itemRepository;
+        this.rentRepository = rentRepository;
     }
 
     public ItemDTO addItem(ItemDTO itemDTO) {
@@ -49,7 +56,7 @@ public class ItemService {
                 return ItemMapper.toComicsDTO((Comics) addedItem);
 
             default:
-                throw new IllegalArgumentException("Invalid item type: " + itemDTO.getItemType());
+                throw new InvalidItemTypeException("Invalid item type: " + itemDTO.getItemType());
         }
     }
 
@@ -57,45 +64,126 @@ public class ItemService {
     public ItemDTO getItemById(ObjectId id) {
         Item item = itemRepository.getItemById(id);
         if (item == null) {
-            throw new NullPointerException("Item not found");
+            throw new ItemNotFoundException("Item with ID: " + id + " not found");
         }
         return ItemMapper.toDTO(item);
     }
 
-//    public List<ItemDTO> getItemsByBasePrice(int basePrice) {
-//        List<Item> items = itemRepository.getItemsByBasePrice(basePrice);
-//        return items.stream()
-//                .map(item -> new ItemDTO(
-//                        item.getBasePrice(),
-//                        item.getItemName(),
-//                        item.isAvailable()
-//                )).collect(Collectors.toList());
-//    }
-//
-//    public List<ItemDTO> getItemsByItemName(String itemName) {
-//        List<Item> items = itemRepository.getItemsByItemName(itemName);
-//        return items.stream()
-//                .map(item -> new ItemDTO(
-//                        item.getBasePrice(),
-//                        item.getItemName(),
-//                        item.isAvailable()
-//                )).collect(Collectors.toList());
-//    }
-//
-//    public List<ItemDTO> getItemsByItemType(String itemType) {
-//        List<Item> items = itemRepository.getItemsByItemType(itemType);
-//        return items.stream()
-//                .map(item -> new ItemDTO(
-//                        item.getBasePrice(),
-//                        item.getItemName(),
-//                        item.isAvailable()
-//                )).collect(Collectors.toList());
-//    }
+    public List<ItemDTO> getItemsByBasePrice(int basePrice) {
+        List<Item> items = itemRepository.getItemsByBasePrice(basePrice);
+        return items.stream()
+                .map(item -> {
+                    if (item instanceof Music) {
+                        Music music = (Music) item;
+                        return new MusicDTO(
+                                music.getId(),
+                                music.getBasePrice(),
+                                music.getItemName(),
+                                music.getGenre(),
+                                music.isVinyl()
+                        );
+                    } else if (item instanceof Movie) {
+                        Movie movie = (Movie) item;
+                        return new MovieDTO(
+                                movie.getId(),
+                                movie.getBasePrice(),
+                                movie.getItemName(),
+                                movie.getMinutes(),
+                                movie.isCasette()
+                        );
+                    } else if (item instanceof Comics) {
+                        Comics comics = (Comics) item;
+                        return new ComicsDTO(
+                                comics.getId(),
+                                comics.getBasePrice(),
+                                comics.getItemName(),
+                                comics.getPageNumber()
+                        );
+                    } else {
+                        throw new InvalidItemTypeException("Unknown item type: " + item.getClass());
+                    }
+                })
+                .collect(Collectors.toList());
+    }
+
+    public List<ItemDTO> getItemsByItemName(String itemName) {
+        List<Item> items = itemRepository.getItemsByItemName(itemName);
+        return items.stream()
+                .map(item -> {
+                    if (item instanceof Music) {
+                        Music music = (Music) item;
+                        return new MusicDTO(
+                                music.getId(),
+                                music.getBasePrice(),
+                                music.getItemName(),
+                                music.getGenre(),
+                                music.isVinyl()
+                        );
+                    } else if (item instanceof Movie) {
+                        Movie movie = (Movie) item;
+                        return new MovieDTO(
+                                movie.getId(),
+                                movie.getBasePrice(),
+                                movie.getItemName(),
+                                movie.getMinutes(),
+                                movie.isCasette()
+                        );
+                    } else if (item instanceof Comics) {
+                        Comics comics = (Comics) item;
+                        return new ComicsDTO(
+                                comics.getId(),
+                                comics.getBasePrice(),
+                                comics.getItemName(),
+                                comics.getPageNumber()
+                        );
+                    } else {
+                        throw new InvalidItemTypeException("Unknown item type: " + item.getClass());
+                    }
+                })
+                .collect(Collectors.toList());
+    }
+
+    public List<ItemDTO> getItemsByItemType(String itemType) {
+        List<Item> items = itemRepository.getItemsByItemType(itemType);
+        return items.stream()
+                .map(item -> {
+                    if (item instanceof Music) {
+                        Music music = (Music) item;
+                        return new MusicDTO(
+                                music.getId(),
+                                music.getBasePrice(),
+                                music.getItemName(),
+                                music.getGenre(),
+                                music.isVinyl()
+                        );
+                    } else if (item instanceof Movie) {
+                        Movie movie = (Movie) item;
+                        return new MovieDTO(
+                                movie.getId(),
+                                movie.getBasePrice(),
+                                movie.getItemName(),
+                                movie.getMinutes(),
+                                movie.isCasette()
+                        );
+                    } else if (item instanceof Comics) {
+                        Comics comics = (Comics) item;
+                        return new ComicsDTO(
+                                comics.getId(),
+                                comics.getBasePrice(),
+                                comics.getItemName(),
+                                comics.getPageNumber()
+                        );
+                    } else {
+                        throw new InvalidItemTypeException("Unknown item type: " + item.getClass());
+                    }
+                })
+                .collect(Collectors.toList());
+    }
 
     public void updateItem(ObjectId id, ItemDTO itemDTO) {
         Item item = itemRepository.getItemById(id);
         if (item == null) {
-            throw new NullPointerException("Item not found");
+            throw new ItemNotFoundException("Item with ID: " + id + " not found");
         }
 
         item.setBasePrice(itemDTO.getBasePrice());
@@ -123,7 +211,7 @@ public class ItemService {
                 break;
 
             default:
-                throw new IllegalArgumentException("Unknown item type: " + itemDTO.getClass().getSimpleName());
+                throw new InvalidItemTypeException("Unknown item type: " + itemDTO.getClass().getSimpleName());
         }
 
         itemRepository.updateItem(item);
@@ -132,7 +220,12 @@ public class ItemService {
     public void removeItem(ObjectId id) {
         Item item = itemRepository.getItemById(id);
         if (item == null) {
-            throw new NullPointerException("Item not found");
+            throw new ItemNotFoundException("Item with ID: " + id + " not found");
+        }
+
+        List<Rent> activeRents = rentRepository.findActiveRentsByItemId(id);
+        if (!activeRents.isEmpty()) {
+            throw new ItemAlreadyRentedException("Item cannot be removed because it is currently rented.");
         }
         itemRepository.removeItem(id);
     }
@@ -140,7 +233,7 @@ public class ItemService {
     public void setAvailable(ObjectId id) {
         Item item = itemRepository.getItemById(id);
         if (item == null) {
-            throw new NullPointerException("Item not found");
+            throw new ItemNotFoundException("Item with ID: " + id + " not found");
         }
         item.setAvailable(true);
         itemRepository.updateItem(item);
@@ -149,7 +242,7 @@ public class ItemService {
     public void setUnavailable(ObjectId id) {
         Item item = itemRepository.getItemById(id);
         if (item == null) {
-            throw new IllegalArgumentException("Item not found with id: " + id);
+            throw new ItemNotFoundException("Item with ID: " + id + " not found");
         }
         item.setAvailable(false);
         itemRepository.updateItem(item);
